polish implementation, add unit test for planner
