+++ src/frontend/planner_test/tests/testdata/delete.yaml
@@ -1,0 +1,20 @@
+# This file is automatically generated. See `src/frontend/planner_test/README.md` for more information.
+- name: delete with returning
+  sql: |
+    create table t (a int, b int);
+    delete from t returning *, a, a+b;
+  logical_plan: |
+    LogicalProject { exprs: [t.a, t.b, t.a, (t.a + t.b)] }
+    └─LogicalDelete { table: t }
+      └─LogicalScan { table: t, columns: [t.a, t.b, t._row_id] }
+  batch_plan: |
+    BatchExchange { order: [], dist: Single }
+    └─BatchProject { exprs: [t.a, t.b, t.a, (t.a + t.b)] }
+      └─BatchDelete { table: t }
+        └─BatchExchange { order: [], dist: Single }
+          └─BatchScan { table: t, columns: [t.a, t.b, t._row_id], distribution: UpstreamHashShard(t._row_id) }
+- name: insert with returning agg functions, should not run
+  sql: |
+    create table t (a int, b int);
+    delete from t returning sum(a);
+  binder_error: 'Bind error: DELETE should not have agg/window'
