+++ src/frontend/src/binder/select.rs
@@ -16,7 +16,7 @@ use std::fmt::Debug;
 
 use itertools::Itertools;
 use risingwave_common::catalog::{Field, Schema, PG_CATALOG_SCHEMA_NAME};
-use risingwave_common::error::{ErrorCode, Result};
+use risingwave_common::error::{ErrorCode, Result, RwError};
 use risingwave_common::types::DataType;
 use risingwave_sqlparser::ast::{Distinct, Expr, Select, SelectItem};
 
@@ -260,6 +260,31 @@ impl Binder {
             }
         }
         Ok((select_list, aliases))
+    }
+
+    pub fn bind_returning_list(
+        &mut self,
+        returning_items: Vec<SelectItem>,
+    ) -> Result<(Vec<ExprImpl>, Vec<Field>)> {
+        let (returning_list, aliases) = self.bind_select_list(returning_items)?;
+        if returning_list
+            .iter()
+            .any(|expr| expr.has_agg_call() || expr.has_window_function())
+        {
+            return Err(RwError::from(ErrorCode::BindError(
+                "DELETE should not have agg/window".to_string(),
+            )));
+        }
+
+        let fields = returning_list
+            .iter()
+            .zip_eq(aliases.iter())
+            .map(|(s, a)| {
+                let name = a.clone().unwrap_or_else(|| UNNAMED_COLUMN.to_string());
+                Ok(Field::with_name(s.return_type(), name))
+            })
+            .collect::<Result<Vec<Field>>>()?;
+        Ok((returning_list, fields))
     }
 
     /// `bind_get_user_by_id_select` binds a select statement that returns a single user name by id,
