+++ src/frontend/src/planner/delete.rs
@@ -17,7 +17,6 @@ use risingwave_common::error::Result;
 
 use super::Planner;
 use crate::binder::BoundDelete;
-use crate::optimizer::plan_node::generic::Project;
 use crate::optimizer::plan_node::{LogicalDelete, LogicalFilter, LogicalProject};
 use crate::optimizer::property::{Order, RequiredDist};
 use crate::optimizer::{PlanRef, PlanRoot};
@@ -34,20 +33,18 @@ impl Planner {
         let mut plan: PlanRef = LogicalDelete::create(input, delete.table_name.clone(), delete.table_id, returning)?.into();
 
         if returning {
-            let len = delete.returning_list.len();
             plan = LogicalProject::create(plan, delete.returning_list);
-            plan = LogicalProject::with_core_and_schema(
-                Project::with_out_col_idx(plan, 0..len),
-                delete.schema,
-            )
-            .into();
         }
 
         // For delete, frontend will only schedule one task so do not need this to be single.
         let dist = RequiredDist::Any;
         let mut out_fields = FixedBitSet::with_capacity(plan.schema().len());
         out_fields.insert_range(..);
-        let out_names = plan.schema().names();
+        let out_names = if returning {
+            delete.returning_schema.expect("If returning list is not empty, should provide returning schema in BoundDelete.").names()
+        } else {
+            plan.schema().names()
+        };
 
         let root = PlanRoot::new(plan, dist, Order::any(), out_fields, out_names);
         Ok(root)
